Make EVM log tests more deterministic
If the from block isn't specified when requesting EVM logs then the logs
from the latest block will be returned, but the latest block changes
with time if empty blocks are being generated, so these tests were
relying on timing to pass. I've fixed the tests by explicitly providing
the from block.