Override read URL for PlasmaChain production cluster (#211)
Only doing this for the most common code path, override can still be
bypassed if needed.

Can't build `sha3` NPM package with NodeJS 11.9.0 that's now used on
Travis by default. Probably just need to upgrade the `sha3` package
but that's going to require some testing first, so downgrading NodeJS
on Travis for now.