Fix LocalAddress.equals (#305)
Previously the `LocalAddress.bytes` could end up being a `Buffer` or a `Uint8Array` depending on what was passed into the constructor, this caused `LocalAddress.equals` to blow up in some cases because it expects to be dealing with `Buffer`. This fix will ensure `LocalAddress.bytes` is always a `Buffer`, which will hopefully make `LocalAddress.equals` more reliable.