Disable another test that tries to use LoomProvider with /eth endpoint
LoomProvider and the underlying Client have to be configured properly
for eth_call to work (used by web3 for non-mutable calls to contracts),
and this test wasn't doing it. The current LoomProvider needs to be
replaced with one resembling the Truffle HDWallet provider, when that
happens this test can be re-enabled & updated to use that new provider.