Add getEthereumTxHash helper function (#339)
A bunch of the tests that use LoomProvider with the /eth endpoint had to be disabled with the new loom build. This is primarily due to the tests failing to configure the accounts used by the LoomProvider and the underlying Client properly in order for `eth_call` to work (used by web3 for non-mutable calls to contracts), configuring this properly is likely non-trivial so all affected tests have been disabled for now. The current LoomProvider needs to be replaced with something resembling the Truffle HDWallet provider, when that happens these tests can be re-enabled & updated to use that new provider.

Co-authored-by: Deeprom <grooveeeeee@gmail.com>